+++ src/sc-as-network.cpp
@@ -22,7 +22,7 @@ std::string sc::random_id (size_t len) {
 //'
 //' @noRd
 // [[Rcpp::export]]
-Rcpp::CharacterVector rcpp_gen_hash (const int n, const int hash_len)
+Rcpp::CharacterVector rcpp_gen_hash (const int n, const size_t hash_len)
 {
     Rcpp::CharacterVector res (n);
     for (int i = 0; i < n; i++)
