+++ src/sc-as-network.h
@@ -5,4 +5,4 @@ namespace sc {
     std::string random_id (size_t len);
 }
 
-Rcpp::CharacterVector rcpp_gen_hash (const int n, const int hash_len);
+Rcpp::CharacterVector rcpp_gen_hash (const int n, const size_t hash_len);
